+++ src/CPluginVariousStuff.h
@@ -40,7 +40,7 @@ namespace VariousStuffPlugin
 
             const char* GetVersion() const
             {
-                return "1.0.3.0";
+                return "1.0.4.0";
             };
 
             const char* GetName() const