+++ changelog.md
@@ -1,3 +1,10 @@
+VariousStuff Plugin 1.0.4.0 (03.03.2013)
+---------------------------
+Stable Release for CryEngine 3.4.5 (32/64 bit)
+
+New:
+* ValueChannel/Signaler Flownode (Tutorial 4)
+
 VariousStuff Plugin 1.0.3.0 (02.03.2013)
 ---------------------------
 Stable Release for CryEngine 3.4.5 (32/64 bit)